% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaoQarea.R
\name{RaoQarea}
\alias{RaoQarea}
\title{RaoQarea}
\usage{
RaoQarea(x, y)
}
\arguments{
\item{x}{A Rasterlayer}

\item{y}{A shapefile where the Raster has to be cropped}
}
\value{
The packege returns the Rao's Q values of the whole cropped area, without making use of the moving window approach
}
\description{
This package returns the Rao's Q values (using an Euclidean distance) for a considered area (raster cropeed by a shapefile) without making use of the moving windows approach
}
